directory list is up in the right place and responsive
